+++ packages/nimble-components/src/table/testing/table.pageobject.ts
@@ -528,7 +528,7 @@ export class TablePageObject<T extends TableRecord> {
      */
     public dragSizeColumnByRightDivider(
         columnIndex: number,
-        deltas: number[]
+        deltas: readonly number[]
     ): void {
         const divider = this.getColumnRightDivider(columnIndex);
         if (!divider) {
@@ -563,7 +563,7 @@ export class TablePageObject<T extends TableRecord> {
      */
     public dragSizeColumnByLeftDivider(
         columnIndex: number,
-        deltas: number[]
+        deltas: readonly number[]
     ): void {
         const divider = this.getColumnLeftDivider(columnIndex);
         if (!divider) {