+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Thu, 04 Jan 2024 19:22:21 GMT and should not be manually modified.
+This log was last generated on Thu, 04 Jan 2024 21:21:59 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 20.17.0
+
+Thu, 04 Jan 2024 21:21:59 GMT
+
+### Minor changes
+
+- - Update all tests to use `parameterizeNamedList` and delete deprecated `getSpecTypeByNamedList` function.
+- Update functions in the table page object and table validator to have `readonly` arrays as the parameters. ([ni/nimble@85b29ec](https://github.com/ni/nimble/commit/85b29ec1836a8bb00b51797c42e4f4cfa058accb))
 
 ## 20.16.6